moves files to more suitable location
